backend-central/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.2
# Communication
websockets==12.0
# HTTP client for external APIs (parking fees, etc.)
requests==2.31.0
# For go2rtc.yaml parsing
pyyaml==6.0.1
# For file upload support
python-multipart==0.0.6